Change from click to argparse
The only external dependency of run-iocsh was click that could be
removed in favor of argparse. Keeping unnecessary external dependencies
potentially add more maintenance, security issues, and disk space.
This change removes click for good.

In python 3.6 argparse API is not prepared for failing tests. We should
update tests when we update the requirements to python 3.9.